move subtests to separate env
