+++ pyproject.toml
@@ -56,6 +56,7 @@ packages = ["pytest_tap"]
 
 [tool.pytest.ini_options]
 pythonpath = [".", "src"]
+markers = ["subtests"]
 
 [tool.uv.sources]
 pytest-tap = { workspace = true }
