+++ tox.ini
@@ -1,17 +1,22 @@
 [tox]
 envlist =
+    subtests
     coverage
 
 [testenv]
 deps =
     pytest
+commands = pytest -m 'not subtests' --tap-combined {posargs}
+
+[testenv:subtests]
+deps =
+    pytest
     pytest-subtests
-commands = pytest --tap-combined {posargs}
+commands = pytest -m 'subtests' --tap-combined {posargs}
 
 [testenv:coverage]
 deps =
     pytest
     pytest-cov
-    pytest-subtests
 commands =
     pytest --cov=pytest_tap --cov-report xml --cov-report term
